V1.2 Oauth implementation (#4)
* Oauth2 support initial commit

* Added authorization check for recipes.

* Hotfix for security check on userId. Unable to delete others recipes.

* Exceptions & handlers now have their own package

* Initial tests added

* Exceptions moved...

* Won't create recipes without owners. @Notnull user

* Added /user endpoint for users to retrieve recipes they've submitted

* Refactored exceptions.

+++ src/main/java/com/herokuapp/receptio/service/RecipeServiceImpl.java
@@ -2,6 +2,8 @@ package com.herokuapp.receptio.service;
 
 import com.herokuapp.receptio.model.Recipe;
 import com.herokuapp.receptio.repository.RecipeRepository;
+import com.herokuapp.receptio.exception.InvalidAuthorizationException;
+import com.herokuapp.receptio.exception.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -21,16 +23,18 @@ public class RecipeServiceImpl implements RecipeService {
     private final RecipeRepository recipeRepository;
 
     @Override
-    public List<Recipe> findAll() {
-        return recipeRepository.findAll();
+    public List<Recipe> findAll(int limit) {
+        return recipeRepository.findAll(limit);
     }
 
     @Override
-    public Recipe findById(Long idRecipe) {
+    public Recipe findById(Long idRecipe) throws ResourceNotFoundException {
         Recipe recipe = null;
         Optional<Recipe> result = recipeRepository.findById(idRecipe);
         if (result.isPresent()) {
             recipe = result.get();
+        } else {
+            throw new ResourceNotFoundException("Did not find recipe with ID: " + idRecipe);
         }
         return recipe;
     }
@@ -41,19 +45,56 @@ public class RecipeServiceImpl implements RecipeService {
     }
 
     @Override
-    public void deleteById(Long idRecipe) {
-        recipeRepository.deleteById(idRecipe);
+    public Recipe update(Recipe recipe) throws ResourceNotFoundException, InvalidAuthorizationException {
+        String existingRecipeUser = findById(recipe.getId()).getUser();
+        String user = recipe.getUser();
+
+        if (!user.equals(existingRecipeUser)) {
+            throw new InvalidAuthorizationException();
+        } else {
+            return recipeRepository.save(recipe);
+        }
     }
 
     @Override
-    public List<Recipe> findRecipesByNameContaining(String searchString) {
-        return recipeRepository.findRecipesByNameContainsIgnoreCase(searchString);
+    public void deleteById(Long idRecipe, String user) throws ResourceNotFoundException, InvalidAuthorizationException {
+        try {
+            Recipe tempRecipe = findById(idRecipe);
+            if (tempRecipe.getUser().equals(user)) {
+                recipeRepository.deleteById(idRecipe);
+            } else {
+                throw new InvalidAuthorizationException();
+            }
+        } catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException("No resource found with ID: " + idRecipe);
+        }
+
     }
 
     @Override
-    public List<Recipe> findRecipesByIngredientsContains(String ingredients) {
+    public List<Recipe> findRecipesByName(String recipeName, int limit) {
+        logger.info("Attempting to find recipes with searchString: " + recipeName);
+        return recipeRepository.findRecipesByNameContainsIgnoreCase(recipeName, limit);
+    }
+
+    @Override
+    public List<Recipe> findRecipesByIngredients(String ingredients, int limit) {
         List<String> ingredientList = Stream.of(ingredients.split(",")).map(String::trim).collect(Collectors.toList());
-        return recipeRepository.findAllByIngredientsNameContainsIgnoreCase(ingredientList);
+        logger.info("Attempting to find recipes with ingredients: " + Arrays.toString(ingredientList.toArray()));
+        return recipeRepository.findAllByIngredientsNameContainsAllIgnoreCase(ingredientList, limit);
+    }
+
+    @Override
+    public List<Recipe> findRecipesByNameAndIngredients(String recipeName, String ingredients, int limit) {
+        List<String> ingredientList = Stream.of(ingredients.split(",")).map(String::trim).collect(Collectors.toList());
+        logger.info("Attempting to find recipes with ingredients: " + Arrays.toString(ingredientList.toArray()));
+        logger.info("Attempting to find recipes with recipeName: " + recipeName);
+        return recipeRepository.findRecipesByNameContainsAndIngredientsNameContainingAllIgnoreCase(recipeName, ingredientList, limit);
+    }
+
+    @Override
+    public List<Recipe> findRecipesByUser(String userId) {
+        return recipeRepository.findAllByUser(userId);
     }
 
 }
